requirements.txt
pyyaml==6.0
paramik==2.11.0